Rename examples to testbed.
