+++ testbed/OpenTracing.Testbed/ClientServer/Server.cs
@@ -5,7 +5,7 @@ using System.Threading;
 using OpenTracing.Propagation;
 using OpenTracing.Tag;
 
-namespace OpenTracing.Examples.ClientServer
+namespace OpenTracing.Testbed.ClientServer
 {
     public class Server
     {
